+++ 04_cronJob.tf
@@ -404,7 +404,7 @@ resource "kubernetes_cron_job_v1" "cronJob" {
                 }
 
                 dynamic "volume_mount" {
-                  for_each = var.applicationConfig.secretVolumes
+                  for_each = local.secretVolumeMounts
                   content {
                     mount_path = volume_mount.value.path
                     name       = kubernetes_secret_v1.secretVolume[volume_mount.key].metadata.0.name
@@ -412,10 +412,28 @@ resource "kubernetes_cron_job_v1" "cronJob" {
                 }
 
                 dynamic "volume_mount" {
-                  for_each = var.applicationConfig.configVolumes
+                  for_each = local.secretVolumeSubpathMounts
+                  content {
+                    mount_path = volume_mount.value.path
+                    name       = kubernetes_secret_v1.secretVolume[volume_mount.value.key].metadata.0.name
+                    sub_path = volume_mount.value.file
+                  }
+                }
+
+                dynamic "volume_mount" {
+                  for_each = local.configVolumeMounts
                   content {
                     mount_path = volume_mount.value.path
                     name       = kubernetes_config_map_v1.configVolume[volume_mount.key].metadata.0.name
+                  }
+                }
+
+                dynamic "volume_mount" {
+                  for_each = local.configVolumeSubpathMounts
+                  content {
+                    mount_path = volume_mount.value.path
+                    name       = kubernetes_config_map_v1.configVolume[volume_mount.value.key].metadata.0.name
+                    sub_path = volume_mount.value.file
                   }
                 }
 
@@ -650,7 +668,7 @@ resource "kubernetes_cron_job_v1" "cronJob" {
                 }
 
                 dynamic "volume_mount" {
-                  for_each = var.applicationConfig.secretVolumes
+                  for_each = local.secretVolumeMounts
                   content {
                     mount_path = volume_mount.value.path
                     name       = kubernetes_secret_v1.secretVolume[volume_mount.key].metadata.0.name
@@ -658,10 +676,28 @@ resource "kubernetes_cron_job_v1" "cronJob" {
                 }
 
                 dynamic "volume_mount" {
-                  for_each = var.applicationConfig.configVolumes
+                  for_each = local.secretVolumeSubpathMounts
+                  content {
+                    mount_path = volume_mount.value.path
+                    name       = kubernetes_secret_v1.secretVolume[volume_mount.value.key].metadata.0.name
+                    sub_path = volume_mount.value.file
+                  }
+                }
+
+                dynamic "volume_mount" {
+                  for_each = local.configVolumeMounts
                   content {
                     mount_path = volume_mount.value.path
                     name       = kubernetes_config_map_v1.configVolume[volume_mount.key].metadata.0.name
+                  }
+                }
+
+                dynamic "volume_mount" {
+                  for_each = local.configVolumeSubpathMounts
+                  content {
+                    mount_path = volume_mount.value.path
+                    name       = kubernetes_config_map_v1.configVolume[volume_mount.value.key].metadata.0.name
+                    sub_path = volume_mount.value.file
                   }
                 }