+++ 22_locals.tf
@@ -12,3 +12,13 @@ locals {
 
 
 }
+
+locals {
+  configVolumeMounts = { for k,v in var.applicationConfig.configVolumes: k => v if v.enableSubpathMount == false}
+  tmpConfigVolumeSubpathMounts = { for k,v in var.applicationConfig.configVolumes: k => merge(v, { key = k}) if v.enableSubpathMount == true }
+  configVolumeSubpathMounts = merge([ for k,v in local.tmpConfigVolumeSubpathMounts: {for f, z in merge(v.data, v.binaryData): "${k}-${f}" => { key = v.key, file = f, path = "${trimsuffix(v.path, "/")}/${f}"}} ]...)
+
+  secretVolumeMounts = { for k,v in var.applicationConfig.secretVolumes: k => v if v.enableSubpathMount == false}
+  tmpSecretVolumeSubpathMounts = { for k,v in var.applicationConfig.secretVolumes: k => merge(v, { key = k}) if v.enableSubpathMount == true }
+  secretVolumeSubpathMounts = merge([ for k,v in local.tmpSecretVolumeSubpathMounts: {for f, z in merge(v.data, v.binaryData): "${k}-${f}" => { key = v.key, file = f, path = "${trimsuffix(v.path, "/")}/${f}"}} ]...)
+}